Show win11 native pin dialog as centered child
The current implementation worked for earlier Win10 builds, but fails for Win11 (at least 24H2). The pin dialog is no longer shown by the requesting process, but a separate CredentialUIBroker process. Therefor, it is not enough to listen for new windows of the unblocker process, but all processes need to be observed.